feat: update header icon
