--- @kiva/kv-components/vue/KvKivaPartnerHeader.vue
@@ -1,44 +1,0 @@
-<!-- eslint-disable max-len -->
-<template>
-	<nav class="tw-bg-primary">
-		<div
-			class="header__full tw-grid xl:tw-gap-x-4 tw-items-center tw-mx-auto tw-px-2.5 md:tw-px-4 lg:tw-px-8"
-			style="height: 96px; max-width: 1200px;"
-		>
-			<div class="header__full__logo">
-				<!-- kivaPartner logo SVG -->
-				<svg
-					width="163"
-					height="48"
-					viewBox="0 0 55 24"
-					fill="none"
-					xmlns="http://www.w3.org/2000/svg"
-				>
-					<path
-						d="M4.56965 0.399536H0V23.1675H4.57195V0.399536H4.56965ZM16.5793 23.1675H21.1513V7.69252H16.5793V23.1675ZM28.2721 23.1675H33.533L38.7938 7.69252H34.2219C34.2219 7.69252 31.7786 16.083 31.0346 18.7523H30.9818C30.2378 16.083 27.7394 7.69252 27.7394 7.69252H22.9034L28.2721 23.1675ZM54.7371 7.69252H50.3259V9.81649C49.5819 8.18284 47.8367 7.25719 45.7539 7.25719C42.0362 7.25719 39.0579 10.3618 39.0579 15.43C39.0579 20.4982 41.9283 23.6005 45.5404 23.6005C48.0938 23.6005 49.5796 22.3472 50.3764 20.9312V23.1675H54.7876L54.7371 7.69252ZM50.2202 15.8103C50.2202 18.6446 48.521 20.0056 46.8194 20.0056C44.9594 20.0056 43.7378 18.0443 43.7378 15.4231C43.7378 12.802 45.0122 10.8407 46.9801 10.8407C48.5738 10.8407 50.2202 12.2017 50.2202 15.0359V15.8103ZM5.73847 15.0474C13.0729 15.0474 14.7744 8.78314 14.7744 7.69252H14.1361C6.80166 7.69252 5.10239 13.959 5.10239 15.0474H5.73847ZM5.10239 15.595C5.10239 16.7406 6.37684 23.1148 14.2417 23.1148H14.8801C14.8801 21.9692 13.6056 15.595 5.73847 15.595H5.10239ZM21.6358 3.45145C21.6358 1.97819 20.5198 0.727182 19.0846 0.727182H18.9239C18.2242 0.7087 17.5458 0.968118 17.0377 1.44844C16.5296 1.92875 16.2334 2.59068 16.2142 3.28878V3.45145C16.2142 4.92471 17.3302 6.17572 18.7654 6.17572H18.9239C19.6239 6.19482 20.3029 5.93568 20.8115 5.4553C21.3201 4.97492 21.6166 4.31264 21.6358 3.61413V3.45145Z"
-						fill="#2AA967"
-					/>
-				</svg>
-			</div>
-
-			<div class="header__full__right">
-				<slot></slot>
-			</div>
-		</div>
-	</nav>
-</template>
-
-<style scoped>
-.header__full {
-  grid-template-areas: 'logo right';
-  grid-template-columns: 1fr auto;
-}
-
-.header__full__logo {
-  grid-area: logo;
-}
-
-.header__full__right {
-  grid-area: right;
-}
-</style>
