--- @kiva/kv-components/vue/stories/KvKivaPartnerHeader.stories.js
@@ -1,20 +1,0 @@
-import KvKivaPartnerHeader from '../KvKivaPartnerHeader.vue';
-
-export default {
-	title: 'KvKivaPartnerHeader',
-	component: KvKivaPartnerHeader,
-};
-
-const story = (args) => {
-	const template = (_args, { argTypes }) => ({
-		props: Object.keys(argTypes),
-		components: { KvKivaPartnerHeader },
-		template: `
-			<kv-kiva-partner-header />
-		`,
-	});
-	template.args = args;
-	return template;
-};
-
-export const Default = story();
